v2.4.4.5 -> v2.4.4.6
  - Johannes Erdfelt: OHCI hash-chain corruption fix, USB updates
  - Richard Henderson, Ivan Kokshaysky: alpha PCI iommu fixes
  - Tim Waugh: parport changelogs and printk levels
  - Andrew Morton: vmalloc off-by-one (overly sensitive) test
  - Al Viro: VFS layer cleanups
  - Cort Dougan: PPC updates (big bootloader re-org)
  - Alan Cox: more merges, remove phillips camera conversion code
  - Andrea Arkangeli: alpha fixups
  - OGAWA Hirofumi: big-sector support with FAT
  - Neil Brown: more md fixes

+++ fs/freevxfs/vxfs_bmap.c
@@ -169,7 +169,7 @@ vxfs_bmap_indir(struct inode *ip, long indir, int size, long block)
 			struct vxfs_typed_dev4	*typ4 = (struct vxfs_typed_dev4 *)typ;
 			printk(KERN_INFO "\n\nTYPED_DEV4 detected!\n");
 			printk(KERN_INFO "block: %Ld\tsize: %Ld\tdev: %d\n",
-					typ4->vd4_block, typ4->vd4_size, typ4->vd4_dev);
+			       (long long)typ4->vd4_block, (long long)typ4->vd4_size, typ4->vd4_dev);
 			goto fail;
 		}
 		default:
@@ -228,7 +228,7 @@ vxfs_bmap_typed(struct inode *ip, long iblock)
 			struct vxfs_typed_dev4	*typ4 = (struct vxfs_typed_dev4 *)typ;
 			printk(KERN_INFO "\n\nTYPED_DEV4 detected!\n");
 			printk(KERN_INFO "block: %Ld\tsize: %Ld\tdev: %d\n",
-					typ4->vd4_block, typ4->vd4_size, typ4->vd4_dev);
+			       (long long)typ4->vd4_block, (long long)typ4->vd4_size, typ4->vd4_dev);
 			return 0;
 		}
 		default:
